workers/requirements.txt
celery==5.3.6
redis==5.0.1
beautifulsoup4==4.12.3
dnspython==2.6.1
playwright==1.42.0
exifread==3.0.0
google-cloud-firestore==2.14.0
python-dotenv==1.0.1
pydantic==2.6.3
httpx==0.27.0
rich==13.7.1
